Fix ios build.
