+++ prometheus/process_collector_darwin.go
@@ -11,6 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//go:build darwin && !ios
+
 package prometheus
 
 import (
