+++ prometheus/process_collector_mem_nocgo_darwin.go
@@ -11,24 +11,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build darwin && cgo
+//go:build darwin && !ios && !cgo
 
 package prometheus
 
-/*
-int get_memory_info(unsigned long long *rss, unsigned long long *vs);
-*/
-import "C"
-import "fmt"
-
 func getMemory() (*memoryInfo, error) {
-	var rss, vsize C.ulonglong
-
-	if err := C.get_memory_info(&rss, &vsize); err != 0 {
-		return nil, fmt.Errorf("task_info() failed with 0x%x", int(err))
-	}
-
-	return &memoryInfo{vsize: uint64(vsize), rss: uint64(rss)}, nil
+	return nil, notImplementedErr
 }
 
 // describe returns all descriptions of the collector for Darwin.
@@ -41,10 +29,10 @@ func (c *processCollector) describe(ch chan<- *Desc) {
 	ch <- c.maxFDs
 	ch <- c.maxVsize
 	ch <- c.startTime
-	ch <- c.rss
-	ch <- c.vsize
 
 	/* the process could be collected but not implemented yet
+	ch <- c.rss
+	ch <- c.vsize
 	ch <- c.inBytes
 	ch <- c.outBytes
 	*/
